move IncidentResponse functions into ChatOps
